+++ packages/dnb-ui-lib/src/style/themes/theme-ui/prism/dnb-prism-theme.js
@@ -77,7 +77,7 @@ const prismTheme = {
     {
       types: ['selector', 'doctype'],
       style: {
-        color: 'var(--color-sea-green-alt-30)',
+        color: 'var(--color-sea-green-30)',
         fontStyle: 'italic'
       }
     },
@@ -96,7 +96,7 @@ const prismTheme = {
     {
       types: ['tag'],
       style: {
-        color: 'var(--color-sea-green-alt)'
+        color: 'var(--color-sea-green)'
       }
     },
     {